[DISCO-3016] refactor: move top level util modules to the utils folder [load test: (skip)]
