+++ merino/providers/__init__.py
@@ -4,7 +4,7 @@ import asyncio
 import logging
 from timeit import default_timer as timer
 
-from merino import metrics
+from merino.utils import metrics
 from merino.config import settings
 from merino.providers.base import BaseProvider
 from merino.providers.manager import load_providers
